bugs fixed, 거래 프로세스 수정

+++ post/src/main/java/post/web/dto/PostsResponseDto.java
@@ -18,7 +18,7 @@ public class PostsResponseDto {
 	private Long id ;
 	private String title ;
 	private int sell_state;
-	private long price;
+	private float price;
 	
 	private LocalDateTime createdDate ;
